First Commit before changing to use images not pdfs. Editpdf does not work correctly

+++ app/src/main/java/com/ganadoro/pile/domain/models/ImageCropData.kt
@@ -1,8 +1,15 @@
 package com.ganadoro.pile.domain.models
 
+/**
+ * Represents the cropping parameters for an image.
+ * @property x The x-coordinate of the top-left corner of the crop rectangle.
+ * @property y The y-coordinate of the top-left corner of the crop rectangle.
+ * @property width The width of the crop rectangle.
+ * @property height The height of the crop rectangle.
+ */
 data class ImageCropData(
-    val x: Float,
-    val y: Float,
-    val width: Float,
-    val height: Float
+    val x:      Int,
+    val y:      Int,
+    val width:  Int,
+    val height: Int
 )
